feat: move sync status to the settings page
